- connected See all students button to /classrooms/:id/students - restructured how files are stored in TeacherDashboard - restructured router to support nesting teacher dashboard components
