+++ packages/app/src/pages/TeacherDashboard/index.tsx
@@ -1,8 +1,7 @@
 export * from "./AddClassroom";
-export * from "./TeacherDashboardMenu/ClassOverview";
-export * from "./MyClassrooms";
-export * from "./StudentSelect";
 export * from "./AddClassroomLanguage";
 export * from "./AddStudents";
+export * from "./ClassOverview";
 export * from "./InviteCaregivers";
-export * from "./TeacherDashboardMenu/ClassOverview";
+export * from "./MyClassrooms";
+export * from "./StudentSelect";
